chore: clean up types and fix wording
